Add the ability to store queries
This commit adds a dropdown that allows options to be selected and
added. The options correspond to key/value pairs in localstorage,
where the value is query DSL. I tried to iron out the kinks with
managing state between react state and localstorage but there may
still be a bug or two. Something to watch out for.